refactoring project structure to support more advanced games
hopefully this new project structure will be useful for the addition of
more advanced games
