+++ games/three_card_monte/__init__.py
@@ -1,0 +1,2 @@
+from .ui import MonteView
+from .engine import deal_winning_card
