+++ games/three_card_monte/engine.py
@@ -1,0 +1,5 @@
+import random
+
+def deal_winning_card() -> str:
+    """Deals the winning card for Three Card Monte."""
+    return random.choice(["A", "B", "C"])
